requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.142
botocore==1.34.142
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.7
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
multidict==6.0.5
openai==0.27.7
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pytesseract==0.3.10
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
